third_party/WebKit/LayoutTests/web-animations-api/animations-responsive-to-color-change-expected.txt

This is a testharness.js-based test.
FAIL Colors responsive to style changes assert_equals: expected "rgb(0, 34, 102)" but got "rgb(102, 34, 0)"
Harness: the test ran to completion.
